SIGARNV-2693 - Busqueda de Baremos correción de filtros para los importes y datos correctos de Baremos.

+++ siga-core/src/main/java/org/itcgae/siga/DTO/scs/BaremosRequestItem.java
@@ -11,7 +11,7 @@ public class BaremosRequestItem {
 	String nDias;
 	String baremo;
 	String dias;
-	Double disponibilidad;
+	Short disponibilidad;
 	Short numMinimoSimple;
 	Short simpleOImporteIndividual;
 	Short naPartir;
